Add files via upload
Moved run() from Panda and Ant to superclass. Both Panda and Ant have the method run. This method is exactly the same in both classes and they both have the same superclass, Animal. We can conclude that all animals will be able to run and as result, we I've moved run to the animal class.